+++ PlanBuild/Blueprints/Components/ToolComponentBase.cs
@@ -1,4 +1,5 @@
-﻿using PlanBuild.ModCompat;
+﻿using Jotunn.Managers;
+using PlanBuild.ModCompat;
 using PlanBuild.Utils;
 using UnityEngine;
 
@@ -283,6 +284,11 @@ namespace PlanBuild.Blueprints.Components
         /// </summary>
         private bool Player_PlacePiece(On.Player.orig_PlacePiece orig, Player self, Piece piece)
         {
+            if (!Config.AllowBlueprintRune.Value && !SynchronizationManager.Instance.PlayerIsAdmin)
+            {
+                MessageHud.instance.ShowMessage(MessageHud.MessageType.Center, "$msg_blueprintrune_disabled");
+                return false;
+            }
             OnPlacePiece(self, piece);
             return false;
         }